+++ utils/callData/identityChangeCalls.ts
@@ -23,8 +23,21 @@ function writeVerifierData(
   };
 }
 
+function updateProfilePicture(
+  nftContractAddress: string,
+  nft_id: string,
+  id: string
+): Call {
+  return {
+    contractAddress: process.env.NEXT_PUBLIC_NFT_PP_VERIFIER as string,
+    entrypoint: "set_native_pp",
+    calldata: [nftContractAddress, parseInt(nft_id), 0, parseInt(id)],
+  };
+}
+
 const identityChangeCalls = {
   writeVerifierData,
+  updateProfilePicture,
 };
 
 export default identityChangeCalls;
