+++ utils/callData/registrationCalls.ts
@@ -99,18 +99,6 @@ function vatTransfer(amount: string): Call {
   };
 }
 
-function updateProfilePicture(
-  nftContractAddress: string,
-  nft_id: string,
-  id: string
-): Call {
-  return {
-    contractAddress: process.env.NEXT_PUBLIC_NFT_PP_VERIFIER as string,
-    entrypoint: "set_native_pp",
-    calldata: [nftContractAddress, parseInt(nft_id), 0, parseInt(id)],
-  };
-}
-
 function renew(
   encodedDomain: string,
   duration: number,
@@ -145,7 +133,6 @@ const registrationCalls = {
   autoRenew,
   buy_discounted,
   vatTransfer,
-  updateProfilePicture,
   renew,
   multiCallRenewal,
 };
